fixed problem with accordion
